% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaCI.R
\name{betaCI}
\alias{betaCI}
\title{Confidence interval of a beta distribution}
\usage{
betaCI(shape1, shape2, CImass = 0.95)
}
\arguments{
\item{shape1}{Shape parameter passed to beta inverse cumulative distribution function qbeta.}

\item{shape2}{Shape parameter passed to beta inverse cumulative distribution function qbeta.}

\item{CImass}{Central probability mass.}
}
\value{
A numeric vector of the two quantiles bounding the central probability mass.
}
\description{
Confidence interval of a beta distribution
}
\examples{
betaCI(6, 4)
betaCI(2, 50, 0.75)
}
