% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDensCurve.R
\name{betaDensCurve}
\alias{betaDensCurve}
\title{Points and densities along a beta distribution}
\usage{
betaDensCurve(shape1, shape2, pmin = 0, pmax = 1, n = 1000)
}
\arguments{
\item{shape1}{Shape parameter passed to \code{dbeta}.}

\item{shape2}{Shape parameter passed to \code{dbeta}.}

\item{pmin}{Minimum value for density sampling.}

\item{pmax}{Maximum value for density sampling.}
}
\value{
A dataframe with columns for beta variate p and its probability density.
}
\description{
Points and densities along a beta distribution
}
