% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaExpect.R
\name{betaExpect}
\alias{betaExpect}
\title{Expectation of a beta distribution}
\usage{
betaExpect(shape1, shape2)
}
\arguments{
\item{shape1}{Shape parameter for beta distribution.}

\item{shape2}{Shape parameter for beta distribution.}
}
\description{
Expectation of a beta distribution
}
