% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaHDIwidthSim.R
\name{betaHDIwidthSim}
\alias{betaHDIwidthSim}
\title{Simulate the highest density interval (HDI) width for a beta posterior.}
\usage{
betaHDIwidthSim(
  genPriorMean,
  genPriorN,
  audPriorMean = 0.5,
  audPriorN = 2,
  HDImass = 0.95,
  minSamples = 20,
  maxSamples = 200
)
}
\arguments{
\item{genPriorMean}{The mean of the generating distribution.}

\item{genPriorN}{The sum of the generating distributions shape parameters.}

\item{audPriorMean}{The mean of the prior for analysis.}

\item{audPriorN}{The sum of the prior's shape parameters or sample size on which the prior is based.}

\item{HDImass}{The desired mass of the highest density interval (HDI).}

\item{minSamples}{Number of samples to begin simulation.}

\item{maxSamples}{Number of samples to end simulation.}
}
\description{
Simulate the highest density interval (HDI) width for a beta posterior.
}
