% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaParamsMeanAndN.R
\name{betaParamsMeanAndN}
\alias{betaParamsMeanAndN}
\title{Beta shape parameters from mean and effective sample size.}
\usage{
betaParamsMeanAndN(mean, n)
}
\arguments{
\item{mean}{Expected value of the beta distribution.}

\item{n}{Sum of shape parameters, i.e. effective sample size.}
}
\description{
Beta shape parameters from mean and effective sample size.
}
