% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_betaArea.R
\name{geom_betaArea}
\alias{geom_betaArea}
\title{Add a shaded area under the beta density curve
Assumes inherited data with columns p and density}
\usage{
geom_betaArea(pmin, pmax)
}
\arguments{
\item{pmin}{Minimum value of p to shade.}

\item{pmax}{Maximum value of p to shade.}
}
\description{
Add a shaded area under the beta density curve
Assumes inherited data with columns p and density
}
