% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_betaExpectation.R
\name{geom_betaExpectation}
\alias{geom_betaExpectation}
\title{Add a vertical line for expected value under beta density curve.}
\usage{
geom_betaExpectation(shape1, shape2, linetype = "dashed")
}
\arguments{
\item{shape1}{Shape parameter passed to dbeta.}

\item{shape2}{Shape parameter passed to dbeta.}

\item{linetype}{Linetype passed to ggplot2.}
}
\description{
Add a vertical line for expected value under beta density curve.
}
