% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdbeta.R
\name{plotdbeta}
\alias{plotdbeta}
\title{Plot a beta density distribution}
\usage{
plotdbeta(shape1, shape2, expectation = TRUE, HDImass = 0.95)
}
\arguments{
\item{shape1}{Shape parameter passed to dbeta.}

\item{shape2}{Shape parameter passed to dbeta.}

\item{expectation}{Include line for expected value.}

\item{HDImass}{Mass of credible interval to shade. NULL to omit.}
}
\description{
Plot a beta density distribution
}
