% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probOfPosterior.R
\name{probOfPosterior}
\alias{probOfPosterior}
\title{Simulate the highest density interval (HDI) width for a beta posterior.}
\usage{
probOfPosterior(
  genPriorMean,
  genPriorN,
  sampleSize,
  audPriorMean = 0.5,
  audPriorN = 2,
  HDImass = 0.95
)
}
\arguments{
\item{genPriorMean}{The mean of the generating distribution.}

\item{genPriorN}{The sum of the generating distributions shape parameters.}

\item{sampleSize}{The number of samples.}

\item{audPriorMean}{The mean of the prior for analysis.}

\item{audPriorN}{The sum of the prior's shape parameters or sample size on which the prior is based.}

\item{HDImass}{The desired mass of the highest density interval (HDI).}
}
\value{
A dataframe with columns n and HDIwidth.
}
\description{
Simulate the highest density interval (HDI) width for a beta posterior.
}
